backend/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.8
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
passlib==1.7.4
PyJWT==2.9.0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
